X.1–X.6: autograd severance + JVP narrowing P0s
audit-5 P0.

X.1 — src/autograd/function_arithmetic.cpp:423-430 (LinearBackward
F16/BF16 upcast):
needs_upcast branch previously wrapped grad_out.tensor().to(F32) /
saved_x.tensor().to(F32) / saved_w.tensor().to(F32) in Variable(...,
false), discarding incoming grad_fn — downstream matmul started a
fresh chain, second-order grads through mixed-precision Linear
silently lost contributions.  Now routes through
tenzor::nn::variable_cast(*, F32) matching the post-compute
narrow-back path at :436-438.

X.2 — src/autograd/vmap.cpp:224-225, 250-251 (MatMulBackward +
shape_passthrough):
Replaced raw tenzor::transpose(result.tensor(), 0, batch_dim) +
Variable(perm_out, requires_grad) with the Variable overload
autograd::transpose(result, 0, batch_dim).  Restores result.grad_fn()
across ~25 vmap-routed ops: Conv1/2/3d, ConvT, BN1/2/3d, LN, GN,
IN1/2/3d, RMSNorm, MultiHeadAttn, Flash/Fused/Flex/NestedAttn,
MaxPool2d, AvgPool2d, AdaptiveAvg/MaxPool2d, RNN/LSTM/GRU cells,
ConstantPad2d, ReflectionPad2d, ReplicationPad2d.

X.3 — src/autograd/jvp_rules.cpp:3741, 3757, 3786
(FlashAttention/Fused/Flex JVP):
float scale = static_cast<float>(attrs.get_float(AttrKey::Scale)) →
double scale; jvp_sdpa_forward already takes double.  V.8 holdouts.

X.4 — src/autograd/jvp_rules.cpp:318 (jvp_clamp):
Dropped static_cast<float> on min_val/max_val; clamp(Tensor, double,
double) overload exists since R.1.

X.5 — src/autograd/function_arithmetic.cpp:155-156, 236-237, 307-308
(complex Mul/Div/MatMul Wirtinger):
Wirtinger paths took conj(saved_b.tensor()) and re-wrapped as
Variable(..., false), dropping any grad_fn on saved_b — second-order
grads through complex Mul/Div/MatMul were zero on the saved-input
branch.

New tenzor::conj(Variable) overload (declared in
include/tenzor/autograd/ops.hpp after view_as_complex; implemented
in src/autograd/ops.cpp).  Routes through existing ConjBackward
(linear: backward(grad) = conj(grad)) with the same wire-up pattern
as view_as_real.

* MulBackward: tenzor::conj(saved_b) / tenzor::conj(saved_a) —
  graph-carrying when has_saved_variables() populates them.
* DivBackward: tenzor::conj(safe_b) / tenzor::conj(saved_a);
  safe_b remains the zero-guarded denominator (intentionally
  non-differentiable at b=0 per original semantics).
* MatMulBackward: tenzor::transpose(saved_b, ...) then
  tenzor::conj(b_t) (and mirror for A) — composes the two
  Variable-level overloads instead of dropping the chain.

X.6 — src/autograd/function_attention.cpp:110, 178, 184, 246-255
(FlashAttention dropout seed truncation):
Forward saved Philox seed as Int64 but backward did
seed_for_replay = static_cast<uint32_t>(seed_cpu.data<int64_t>()[0])
— discarded upper 32 bits, so tenzor::manual_seed past 2^32
produced different forward and replay masks, corrupting dQ/dK/dV.

Fixes:
* composed_attention_backward rng_seed parameter type uint32_t →
  uint64_t; !=0 guard updated.
* Removed inner static_cast<uint64_t>(rng_seed) (now pass-through to
  philox_dropout_mask which already accepts uint64_t).
* seed_for_replay uint32_t → uint64_t; Int64 path drops the
  truncating cast; Int32 path widens uint32 → uint64 so negative
  seeds zero-extend cleanly.

Verified: ninja tenzor_core clean.

+++ src/autograd/function_attention.cpp
@@ -107,7 +107,7 @@ auto composed_attention_backward(const Tensor& dO,
                                  double scale,
                                  bool causal,
                                  double dropout_p = 0.0,
-                                 uint32_t rng_seed = 0u) -> std::vector<Tensor> {
+                                 uint64_t rng_seed = 0ull) -> std::vector<Tensor> {
     // V.8: double scale end-to-end; audit-3 R.6 had widened only the
     // Variable form. dropout_p is also widened to keep parity with the
     // Variable path; the only downstream effect is the threshold used
@@ -175,13 +175,13 @@ auto composed_attention_backward(const Tensor& dO,
     // philox_dropout_mask. The mask op already returns scale=1/(1-p) for
     // kept positions and 0 for dropped, which is exactly the
     // multiplicative correction we need on P (inverted dropout).
-    if (dropout_p > 0.0f && rng_seed != 0u) {
+    if (dropout_p > 0.0f && rng_seed != 0ull) {
         auto P_shape = P.shape();
         std::vector<int64_t> mask_shape(P_shape.begin(), P_shape.end());
         Tensor mask_cpu = philox_dropout_mask(
             mask_shape,
             static_cast<double>(dropout_p),
-            static_cast<uint64_t>(rng_seed),
+            rng_seed,
             /*offset=*/0,
             DType::Float32);
         Tensor mask_dev = mask_cpu.to(P.device()).to(P.dtype());
@@ -243,14 +243,19 @@ auto try_fused_or_compose_backward(const Tensor& dO,
     // FusedAttentionBackward) already hold `double scale_` and were
     // narrowing through the float parameter at every call.
     // Read seed (if any) for the composed-fallback dropout replay.
-    uint32_t seed_for_replay = 0u;
+    // Audit-5 X.6: previously narrowed Int64 → uint32, discarding the upper
+    // 32 bits of the saved Philox seed. When the forward used the full 64
+    // bits (manual_seed past 2^32) the replay produced a different mask
+    // than the forward → corrupted dQ/dK/dV. philox_dropout_mask accepts
+    // uint64_t natively; thread the full width through.
+    uint64_t seed_for_replay = 0ull;
     if (philox_seed.is_valid() && philox_seed.numel() > 0) {
         // Move to CPU to read scalar; tiny single-int copy.
         Tensor seed_cpu = philox_seed.cpu();
         if (seed_cpu.dtype() == DType::Int64) {
-            seed_for_replay = static_cast<uint32_t>(seed_cpu.data<int64_t>()[0]);
+            seed_for_replay = static_cast<uint64_t>(seed_cpu.data<int64_t>()[0]);
         } else if (seed_cpu.dtype() == DType::Int32) {
-            seed_for_replay = static_cast<uint32_t>(seed_cpu.data<int32_t>()[0]);
+            seed_for_replay = static_cast<uint64_t>(static_cast<uint32_t>(seed_cpu.data<int32_t>()[0]));
         }
     }
